Move over more test helpers
