+++ sbtc/src/testing/mod.rs
@@ -1,2 +1,3 @@
 //! All testing utility functions.
+pub mod deposits;
 pub mod regtest;
